dfp/requirements.txt
numpy
easydict
tensorflow-gpu
matplotlib
scipy
